feat(home): move the sound toggle to Home and add an editor link
"Enable all sounds" sat in Settings, a tab away from the volume sliders it
governs. Move it to the Home page next to them, and add a row linking to
the web sound pack editor below it.

Both rows use the label-left / control-right shape. The editor row is
written out rather than reusing Toggler, which is specific to a checkbox
and takes no arbitrary control.
